fix Int

+++ Sources/MachOParser/MachOTypes/Headers/MachHeaderFlag.swift
@@ -36,66 +36,66 @@ public enum MachHeaderFlag: Option {
     case simSupport
     case dylibInCache
 
-    public var value: Int {
+    public var value: Int64 {
         switch self {
         case .noundefs:
-            return Int(MH_NOUNDEFS)
+            return Int64(MH_NOUNDEFS)
         case .incrlink:
-            return Int(MH_INCRLINK)
+            return Int64(MH_INCRLINK)
         case .dyldlink:
-            return Int(MH_DYLDLINK)
+            return Int64(MH_DYLDLINK)
         case .bindatload:
-            return Int(MH_BINDATLOAD)
+            return Int64(MH_BINDATLOAD)
         case .prebound:
-            return Int(MH_PREBOUND)
+            return Int64(MH_PREBOUND)
         case .splitSegs:
-            return Int(MH_SPLIT_SEGS)
+            return Int64(MH_SPLIT_SEGS)
         case .lazyInit:
-            return Int(MH_LAZY_INIT)
+            return Int64(MH_LAZY_INIT)
         case .twolevel:
-            return Int(MH_TWOLEVEL)
+            return Int64(MH_TWOLEVEL)
         case .forceFlat:
-            return Int(MH_FORCE_FLAT)
+            return Int64(MH_FORCE_FLAT)
         case .nomultidefs:
-            return Int(MH_NOMULTIDEFS)
+            return Int64(MH_NOMULTIDEFS)
         case .nofixprebinding:
-            return Int(MH_NOFIXPREBINDING)
+            return Int64(MH_NOFIXPREBINDING)
         case .prebindable:
-            return Int(MH_PREBINDABLE)
+            return Int64(MH_PREBINDABLE)
         case .allmodsbound:
-            return Int(MH_ALLMODSBOUND)
+            return Int64(MH_ALLMODSBOUND)
         case .subsectionsViaSymbols:
-            return Int(MH_SUBSECTIONS_VIA_SYMBOLS)
+            return Int64(MH_SUBSECTIONS_VIA_SYMBOLS)
         case .canonical:
-            return Int(MH_CANONICAL)
+            return Int64(MH_CANONICAL)
         case .weakDefines:
-            return Int(MH_WEAK_DEFINES)
+            return Int64(MH_WEAK_DEFINES)
         case .bindsToWeak:
-            return Int(MH_BINDS_TO_WEAK)
+            return Int64(MH_BINDS_TO_WEAK)
         case .allowStackExecution:
-            return Int(MH_ALLOW_STACK_EXECUTION)
+            return Int64(MH_ALLOW_STACK_EXECUTION)
         case .rootSafe:
-            return Int(MH_ROOT_SAFE)
+            return Int64(MH_ROOT_SAFE)
         case .setuidSafe:
-            return Int(MH_SETUID_SAFE)
+            return Int64(MH_SETUID_SAFE)
         case .noReexportedDylibs:
-            return Int(MH_NO_REEXPORTED_DYLIBS)
+            return Int64(MH_NO_REEXPORTED_DYLIBS)
         case .pie:
-            return Int(MH_PIE)
+            return Int64(MH_PIE)
         case .deadStrippableDylib:
-            return Int(MH_DEAD_STRIPPABLE_DYLIB)
+            return Int64(MH_DEAD_STRIPPABLE_DYLIB)
         case .hasTlvDescriptors:
-            return Int(MH_HAS_TLV_DESCRIPTORS)
+            return Int64(MH_HAS_TLV_DESCRIPTORS)
         case .noHeapExecution:
-            return Int(MH_NO_HEAP_EXECUTION)
+            return Int64(MH_NO_HEAP_EXECUTION)
         case .appExtensionSafe:
-            return Int(MH_APP_EXTENSION_SAFE)
+            return Int64(MH_APP_EXTENSION_SAFE)
         case .nlistOutofsyncWithDyldinfo:
-            return Int(MH_NLIST_OUTOFSYNC_WITH_DYLDINFO)
+            return Int64(MH_NLIST_OUTOFSYNC_WITH_DYLDINFO)
         case .simSupport:
-            return Int(MH_SIM_SUPPORT)
+            return Int64(MH_SIM_SUPPORT)
         case .dylibInCache:
-            return Int(MH_DYLIB_IN_CACHE)
+            return Int64(MH_DYLIB_IN_CACHE)
         }
     }
 }
